Rework MicroStack docs to use Secure Clustering
As of revision 215 in the store, MicroStack contains the Secure
Clustering feature which also introduces the "microstack" command
with {init, launch, add-compute} subcommands and makes the node
role a mandatory CLI argument to be specified when --auto is used.
https://review.opendev.org/#/c/757658/

Passwords are now randomly generated to avoid insecure practices:
https://review.opendev.org/#/c/756567/

This change addresses the above modifications.

Metrics and GPGPU sections are removed since they have not been
validated after an upgrade to Focal/Ussuri/OVN (and have not
had any auto test coverage prior to that).

https://review.opendev.org/#/c/738242/

Metrics specifically were added as an experimental contribution
which has not received proper validation and testing.